sam/s3_csv_aggregated/function/requirements.txt
python-dotenv==1.0.1
boto3==1.35.95
